requirements.txt
lxml==4.6.3
bs4==0.0.1
beautifulsoup4==4.9.3
soupsieve==2.2.1
pip==20.0.2
attrs==20.3.0
idna==2.10
multidict==5.1.0
asyncio==3.4.3
yarl==1.6.3
aiohttp==3.7.4.post0
async-class==0.3.0
certifi==2020.12.5
chardet==4.0.0
urllib3==1.26.4
six==1.15.0
executing==0.5.4
setuptools==44.0.0
python-dotenv==0.17.0